deploymentwatcher: failing test
Test demonstrating a bug with canary deployments when there are multiple
groups. If group A's canary becomes healthy before group B's, the deadline
for the overall deployment will be set to that of group A. When the deployment
is promoted, if group A is done it will not contribute to the next deadline
cutoff. Group B's old deadline will be used instead, which will be in the past
and immediately trigger a deployment progress failure.